refactor: Update Analytics panel
